WIP: create PhraseScorer abstraction + impl ExactPhraseScorer + implem SlopPhraseScorer
