+++ src/query/phrase_query/phrase_scorer.rs
@@ -1,13 +1,14 @@
-use std::cmp::Ordering;
-
-use crate::docset::{DocSet, TERMINATED};
-use crate::fieldnorm::FieldNormReader;
+use crate::docset::DocSet;
 use crate::postings::Postings;
-use crate::query::bm25::Bm25Weight;
-use crate::query::{Intersection, Scorer};
-use crate::{DocId, Score};
+use crate::query::Scorer;
+use crate::DocId;
 
-struct PostingsWithOffset<TPostings> {
+pub trait PhraseScorer: Scorer {
+    fn phrase_count(&self) -> u32;
+    fn phrase_match(&mut self) -> bool;
+    fn phrase_exists(&mut self) -> bool;
+}
+pub(crate) struct PostingsWithOffset<TPostings> {
     offset: u32,
     postings: TPostings,
 }
@@ -43,483 +44,18 @@ impl<TPostings: Postings> DocSet for PostingsWithOffset<TPostings> {
     }
 }
 
-pub struct PhraseScorer<TPostings: Postings> {
-    intersection_docset: Intersection<PostingsWithOffset<TPostings>, PostingsWithOffset<TPostings>>,
-    num_terms: usize,
-    left: Vec<u32>,
-    right: Vec<u32>,
-    phrase_count: u32,
-    fieldnorm_reader: FieldNormReader,
-    similarity_weight: Bm25Weight,
-    scoring_enabled: bool,
-    slop: u32,
-    used_slop: u32,
-}
-
-/// Returns true if and only if the two sorted arrays contain a common element
-fn intersection_exists(left: &[u32], right: &[u32]) -> bool {
-    let mut left_index = 0;
-    let mut right_index = 0;
-    while left_index < left.len() && right_index < right.len() {
-        let left_val = left[left_index];
-        let right_val = right[right_index];
-        match left_val.cmp(&right_val) {
-            Ordering::Less => {
-                left_index += 1;
-            }
-            Ordering::Equal => {
-                return true;
-            }
-            Ordering::Greater => {
-                right_index += 1;
-            }
-        }
-    }
-    false
-}
-
-fn intersection_count(left: &[u32], right: &[u32]) -> usize {
-    let mut left_index = 0;
-    let mut right_index = 0;
-    let mut count = 0;
-    while left_index < left.len() && right_index < right.len() {
-        let left_val = left[left_index];
-        let right_val = right[right_index];
-        match left_val.cmp(&right_val) {
-            Ordering::Less => {
-                left_index += 1;
-            }
-            Ordering::Equal => {
-                count += 1;
-                left_index += 1;
-                right_index += 1;
-            }
-            Ordering::Greater => {
-                right_index += 1;
-            }
-        }
-    }
-    count
-}
-
-/// Intersect twos sorted arrays `left` and `right` and outputs the
-/// resulting array in left.
-///
-/// Returns the length of the intersection
-fn intersection(left: &mut [u32], right: &[u32]) -> usize {
-    let mut left_index = 0;
-    let mut right_index = 0;
-    let mut count = 0;
-    let left_len = left.len();
-    let right_len = right.len();
-    while left_index < left_len && right_index < right_len {
-        let left_val = left[left_index];
-        let right_val = right[right_index];
-        match left_val.cmp(&right_val) {
-            Ordering::Less => {
-                left_index += 1;
-            }
-            Ordering::Equal => {
-                left[count] = left_val;
-                count += 1;
-                left_index += 1;
-                right_index += 1;
-            }
-            Ordering::Greater => {
-                right_index += 1;
-            }
-        }
-    }
-    count
-}
-
-/// Intersect twos sorted arrays `left` and `right` and outputs the
-/// resulting array in left.
-///
-/// Condition for match is that the value stored in left is less than or equal to
-/// the value in right and that the distance to the previous token is lte to the slop.
-///
-/// Returns the length of the intersection
-fn intersection_with_slop(
-    left: &mut [u32],
-    right: &[u32],
-    slop: u32,
-    used_slop: u32,
-) -> (usize, u32) {
-    let mut left_index = 0;
-    let mut right_index = 0;
-    let mut count = 0;
-    let left_len = left.len();
-    let right_len = right.len();
-    let mut used_slop = used_slop;
-    while left_index < left_len && right_index < right_len {
-        let left_val = left[left_index];
-        let right_val = right[right_index];
-
-        // The three conditions are:
-        // left_val < right_slop -> left index increment.
-        // right_slop <= left_val <= right -> find the best match.
-        // left_val > right -> right index increment.
-        let right_slop = if right_val >= slop {
-            right_val - slop
-        } else {
-            0
-        };
-
-        if left_val < right_slop {
-            left_index += 1;
-        } else if right_slop <= left_val && left_val <= right_val {
-            while left_index + 1 < left_len {
-                // there could be a better match
-                let next_left_val = left[left_index + 1];
-                if next_left_val > right_val {
-                    // the next value is outside the range, so current one is the best.
-                    break;
-                }
-                // the next value is better.
-                left_index += 1;
-            }
-            used_slop += right_val - left[left_index];
-            // store the match in left.
-            left[count] = right_val;
-            count += 1;
-            left_index += 1;
-            right_index += 1;
-        } else if left_val > right_val {
-            right_index += 1;
-        }
-    }
-    (count, used_slop)
-}
-
-fn intersection_count_with_slop(
-    left: &[u32],
-    right: &[u32],
-    slop: u32,
-    used_slop: u32,
-) -> (usize, u32) {
-    let mut left_index = 0;
-    let mut right_index = 0;
-    let mut count = 0;
-    let left_len = left.len();
-    let right_len = right.len();
-    let mut used_slop = used_slop;
-    while left_index < left_len && right_index < right_len {
-        let left_val = left[left_index];
-        let right_val = right[right_index];
-        let right_slop = if right_val >= slop {
-            right_val - slop
-        } else {
-            0
-        };
-
-        if left_val < right_slop {
-            left_index += 1;
-        } else if right_slop <= left_val && left_val <= right_val {
-            while left_index + 1 < left_len {
-                let next_left_val = left[left_index + 1];
-                if next_left_val > right_val {
-                    break;
-                }
-                left_index += 1;
-            }
-            used_slop += right_val - left[left_index];
-            count += 1;
-            left_index += 1;
-            right_index += 1;
-        } else if left_val > right_val {
-            right_index += 1;
-        }
-    }
-    (count, used_slop)
-}
-
-fn intersection_exists_with_slop(left: &[u32], right: &[u32], slop: u32) -> bool {
-    let mut left_index = 0;
-    let mut right_index = 0;
-    let left_len = left.len();
-    let right_len = right.len();
-    while left_index < left_len && right_index < right_len {
-        let left_val = left[left_index];
-        let right_val = right[right_index];
-        let right_slop = if right_val >= slop {
-            right_val - slop
-        } else {
-            0
-        };
-
-        if left_val < right_slop {
-            left_index += 1;
-        } else if right_slop <= left_val && left_val <= right_val {
-            return true;
-        } else if left_val > right_val {
-            right_index += 1;
-        }
-    }
-    false
-}
-
-impl<TPostings: Postings> PhraseScorer<TPostings> {
-    pub fn new(
-        term_postings: Vec<(usize, TPostings)>,
-        similarity_weight: Bm25Weight,
-        fieldnorm_reader: FieldNormReader,
-        scoring_enabled: bool,
-        slop: u32,
-    ) -> PhraseScorer<TPostings> {
-        let max_offset = term_postings
-            .iter()
-            .map(|&(offset, _)| offset)
-            .max()
-            .unwrap_or(0);
-        let num_docsets = term_postings.len();
-        let postings_with_offsets = term_postings
-            .into_iter()
-            .map(|(offset, postings)| {
-                PostingsWithOffset::new(postings, (max_offset - offset) as u32)
-            })
-            .collect::<Vec<_>>();
-        let mut scorer = PhraseScorer {
-            intersection_docset: Intersection::new(postings_with_offsets),
-            num_terms: num_docsets,
-            left: Vec::with_capacity(100),
-            right: Vec::with_capacity(100),
-            phrase_count: 0u32,
-            similarity_weight,
-            fieldnorm_reader,
-            scoring_enabled,
-            slop,
-            used_slop: 0u32,
-        };
-        if scorer.doc() != TERMINATED && !scorer.phrase_match() {
-            scorer.advance();
-        }
-        scorer
-    }
-
-    pub fn phrase_count(&self) -> u32 {
-        self.phrase_count
-    }
-
-    fn phrase_match(&mut self) -> bool {
-        if self.scoring_enabled {
-            let count = self.compute_phrase_count();
-            self.phrase_count = count;
-            count > 0u32
-        } else {
-            self.phrase_exists()
-        }
-    }
-
-    fn phrase_exists(&mut self) -> bool {
-        let (intersection_len, _) = self.compute_phrase_match();
-        if self.has_slop() {
-            return intersection_exists_with_slop(
-                &self.left[..intersection_len],
-                &self.right[..],
-                self.slop,
-            );
-        }
-        intersection_exists(&self.left[..intersection_len], &self.right[..])
-    }
-
-    fn compute_phrase_count(&mut self) -> u32 {
-        let (intersection_len, used_slop) = self.compute_phrase_match();
-        if self.has_slop() {
-            let (count, used_slop) = intersection_count_with_slop(
-                &self.left[..intersection_len],
-                &self.right[..],
-                self.slop,
-                used_slop,
-            );
-            self.used_slop = used_slop;
-            return count as u32;
-        }
-        intersection_count(&self.left[..intersection_len], &self.right[..]) as u32
-    }
-
-    fn compute_phrase_match(&mut self) -> (usize, u32) {
-        {
-            self.intersection_docset
-                .docset_mut_specialized(0)
-                .positions(&mut self.left);
-        }
-        let mut intersection_len = (self.left.len(), 0);
-        for i in 1..self.num_terms - 1 {
-            {
-                self.intersection_docset
-                    .docset_mut_specialized(i)
-                    .positions(&mut self.right);
-            }
-            intersection_len = if self.has_slop() {
-                intersection_with_slop(
-                    &mut self.left[..intersection_len.0],
-                    &self.right[..],
-                    self.slop,
-                    intersection_len.1,
-                )
-            } else {
-                (
-                    intersection(&mut self.left[..intersection_len.0], &self.right[..]),
-                    0,
-                )
-            };
-            if intersection_len.0 == 0 {
-                return (0, 0);
-            }
-        }
-        self.intersection_docset
-            .docset_mut_specialized(self.num_terms - 1)
-            .positions(&mut self.right);
-        intersection_len
-    }
-
-    fn has_slop(&self) -> bool {
-        self.slop > 0
-    }
-}
-
-impl<TPostings: Postings> DocSet for PhraseScorer<TPostings> {
-    fn advance(&mut self) -> DocId {
-        loop {
-            let doc = self.intersection_docset.advance();
-            if doc == TERMINATED || self.phrase_match() {
-                return doc;
-            }
-        }
-    }
-
-    fn seek(&mut self, target: DocId) -> DocId {
-        debug_assert!(target >= self.doc());
-        let doc = self.intersection_docset.seek(target);
-        if doc == TERMINATED || self.phrase_match() {
-            return doc;
-        }
-        self.advance()
-    }
-
-    fn doc(&self) -> DocId {
-        self.intersection_docset.doc()
-    }
-
-    fn size_hint(&self) -> u32 {
-        self.intersection_docset.size_hint()
-    }
-}
-
-impl<TPostings: Postings> Scorer for PhraseScorer<TPostings> {
-    fn score(&mut self) -> Score {
-        let doc = self.doc();
-        let fieldnorm_id = self.fieldnorm_reader.fieldnorm_id(doc);
-        if self.has_slop() {
-            // we use slop+1 rather than slop to garantee the resulting ratio is > 0
-            let allowed_slop = (self.slop + 1) as f32 * self.phrase_count as f32;
-            let spare_slop_ratio = 1f32 - self.used_slop as f32 / allowed_slop;
-            return self
-                .similarity_weight
-                .score(fieldnorm_id, self.phrase_count)
-                * spare_slop_ratio;
-        }
-        self.similarity_weight
-            .score(fieldnorm_id, self.phrase_count)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{intersection, intersection_count, intersection_with_slop};
-
-    fn test_intersection_sym(left: &[u32], right: &[u32], expected: &[u32]) {
-        test_intersection_aux(left, right, expected, 0);
-        test_intersection_aux(right, left, expected, 0);
-    }
-
-    fn test_intersection_aux(left: &[u32], right: &[u32], expected: &[u32], slop: u32) {
-        let mut left_vec = Vec::from(left);
-        let left_mut = &mut left_vec[..];
-        if slop == 0 {
-            let left_mut = &mut left_vec[..];
-            assert_eq!(intersection_count(left_mut, right), expected.len());
-            let count = intersection(left_mut, right);
-            assert_eq!(&left_mut[..count], expected);
-            return;
-        }
-        let mut right_vec = Vec::from(right);
-        let right_mut = &mut right_vec[..];
-        let count = intersection_with_slop(left_mut, right_mut, slop, 0);
-        assert_eq!(&left_mut[..count.0], expected);
-    }
-
-    #[test]
-    fn test_intersection() {
-        test_intersection_sym(&[1], &[1], &[1]);
-        test_intersection_sym(&[1], &[2], &[]);
-        test_intersection_sym(&[], &[2], &[]);
-        test_intersection_sym(&[5, 7], &[1, 5, 10, 12], &[5]);
-        test_intersection_sym(&[1, 5, 6, 9, 10, 12], &[6, 8, 9, 12], &[6, 9, 12]);
-    }
-    #[test]
-    fn test_slop() {
-        // The slop is not symmetric. It does not allow for the phrase to be out of order.
-        test_intersection_aux(&[1], &[2], &[2], 1);
-        test_intersection_aux(&[1], &[3], &[], 1);
-        test_intersection_aux(&[1], &[3], &[3], 2);
-        test_intersection_aux(&[], &[2], &[], 100000);
-        test_intersection_aux(&[5, 7, 11], &[1, 5, 10, 12], &[5, 12], 1);
-        test_intersection_aux(&[1, 5, 6, 9, 10, 12], &[6, 8, 9, 12], &[6, 9, 12], 1);
-        test_intersection_aux(&[1, 5, 6, 9, 10, 12], &[6, 8, 9, 12], &[6, 9, 12], 10);
-        test_intersection_aux(&[1, 3, 5], &[2, 4, 6], &[2, 4, 6], 1);
-        test_intersection_aux(&[1, 3, 5], &[2, 4, 6], &[], 0);
-    }
-
-    fn test_merge(
-        left: &[u32],
-        right: &[u32],
-        expected_left: &[u32],
-        expected_used: u32,
-        slop: u32,
-    ) {
-        let mut left_vec = Vec::from(left);
-        let left_mut = &mut left_vec[..];
-        let mut right_vec = Vec::from(right);
-        let right_mut = &mut right_vec[..];
-        let count = intersection_with_slop(left_mut, right_mut, slop, 0);
-        assert_eq!(&left_mut[..count.0], expected_left);
-        assert_eq!(count.1, expected_used);
-    }
-
-    #[test]
-    fn test_merge_slop() {
-        test_merge(&[1, 2], &[1], &[1], 0, 1);
-        test_merge(&[3], &[4], &[4], 1, 2);
-        test_merge(&[3], &[4], &[4], 1, 2);
-        test_merge(&[1, 5, 6, 9, 10, 12], &[6, 8, 9, 12], &[6, 9, 12], 0, 0);
-        test_merge(&[1, 5, 6], &[2, 6, 7], &[2, 6], 1, 10); // we don't answer 2,6,7
-    }
-}
-
-#[cfg(all(test, feature = "unstable"))]
-mod bench {
-
-    use test::Bencher;
-
-    use super::{intersection, intersection_count};
-
-    #[bench]
-    fn bench_intersection_short(b: &mut Bencher) {
-        b.iter(|| {
-            let mut left = [1, 5, 10, 12];
-            let right = [5, 7];
-            intersection(&mut left, &right);
-        });
-    }
-
-    #[bench]
-    fn bench_intersection_count_short(b: &mut Bencher) {
-        b.iter(|| {
-            let left = [1, 5, 10, 12];
-            let right = [5, 7];
-            intersection_count(&left, &right);
-        });
-    }
+/// In order to check if terms are adjacent, we offset the terms' positions in the field relatively
+/// to their position in the phrase. It then becomes a list intersection problem.
+pub(crate) fn get_postings_with_offset<TPostings: Postings>(
+    term_postings: Vec<(usize, TPostings)>,
+) -> Vec<PostingsWithOffset<TPostings>> {
+    let max_offset = term_postings
+        .iter()
+        .map(|&(offset, _)| offset)
+        .max()
+        .unwrap_or(0);
+    term_postings
+        .into_iter()
+        .map(|(offset, postings)| PostingsWithOffset::new(postings, (max_offset - offset) as u32))
+        .collect::<Vec<_>>()
 }
